+++ notifiers/__init__.py
@@ -2,3 +2,4 @@ from .console import ConsoleNotifier
 from .slack import SlackNotifier
 from .email import EmailNotifier
 from .influxdb import InfluxDBNotifier
+from .influxdb_v2 import InfluxDBV2Notifier